requirements.txt
# conda create -n "expertcourse" python=3.10.13
# pip install -r "requirements_windows.txt"
jupyter==1.0.0
matplotlib==3.8.2
numpy==1.26.4
sympy==1.12
pandas==2.1.4
dwave-ocean-sdk==6.9.0
scipy==1.11.4
networkx==3.2.1
openpyxl==3.1.2
plotly==5.18.0
scikit-learn==1.3.2
seaborn==0.13.0
# .\dadk_light.tar.bz2